fix: Refactor Router, Directory and files
