+++ client/src/appStore/slices/alertes.slice.js
@@ -1,63 +1,37 @@
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { createSlice } from '@reduxjs/toolkit';
 import axios from 'axios';
 
+export const getAlertes = () => async (dispatch) => {
+  dispatch(getAlertesStart());
+  try {
+    const response = await axios.get('http://localhost:8000/api/alertes');
+    dispatch(getAlertesSuccess(response.data));
+  } catch (error) {
+    dispatch(getAlertesFailure(error.message));
+  }
+};
+
 const alerteSlice = createSlice({
-    name: 'alertes',
-    initialState: {
-        loading: false,
-        error: null,
-        data: [],
+  name: 'alertes',
+  initialState: {
+    loading: false,
+    error: null,
+    data: [],
+  },
+  reducers: {
+    getAlertesStart: (state) => {
+      state.loading = true;
     },
-    reducers: {
-        updateAlertes(state, action) {
-            state.data = action.payload;
-        },
-
-        getAlertes(state) {
-            state.loading = true;
-            axios.get('http://localhost:3000/api/alertes')
-                .then(response => {
-                    state.loading = false;
-                    state.data = response.data;
-                })
-                .catch(error => {
-                    state.loading = false;
-                    state.error = error.message;
-                });
-        },
-
-        addAlerte(state, action) {
-            state.loading = true;
-            axios.post('http://localhost:3000/api/alertes', action.payload)
-                .then(response => {
-                    state.loading = false;
-                    state.data.push(response.data);
-                })
-                .catch(error => {
-                    state.loading = false;
-                    state.error = error.message;
-                });
-        },
-
-        removeAlerte(state, action) {
-            state.loading = true;
-            const alerteIdToRemove = action.payload;
-            axios.delete(`http://localhost:3000/api/alertes/${alerteIdToRemove}`)
-                .then(() => {
-                    state.loading = false;
-                    state.data = state.data.filter(alerte => alerte.id !== alerteIdToRemove);
-                })
-                .catch(error => {
-                    state.loading = false;
-                    state.error = error.message;
-                });
-        },
-
-        showAlertes(state) {
-            console.log(state.data);
-        },
-    }
+    getAlertesSuccess: (state, action) => {
+      state.loading = false;
+      state.data = action.payload;
+    },
+    getAlertesFailure: (state, action) => {
+      state.loading = false;
+      state.error = action.payload;
+    },
+  },
 });
 
-export const { updateAlertes, getAlertes, addAlerte, removeAlerte, showAlertes } = alerteSlice.actions;
+export const { getAlertesStart, getAlertesSuccess, getAlertesFailure } = alerteSlice.actions;
 export default alerteSlice.reducer;
