+++ client/src/composants/404/Page404.component.jsx
@@ -1,0 +1,7 @@
+export const Page404 = () => {
+    return (
+        <>
+        
+        </>
+    )
+}
