+++ client/src/composants/Billetterie/Billetterie.component.jsx
@@ -1,0 +1,6 @@
+export const Billetterie = () => {
+    return(
+        <>
+        </>
+    )
+}
