--- client/src/composants/Cards/Large/ArtistCards/artistCard.style.css
@@ -1,12 +1,0 @@
-.artistes-card-title {
-    color: var(--secondary-orange);
-  }
-  
-  .artistes-card-gradient {
-      background-image: linear-gradient(to right, var(--main), var(--secondary-yellow), var(--secondary-orange));
-  }
-  
-
-  .artistes-card-name {
-    color: var(--white-100);
-  }
