--- client/src/composants/Cards/Large/HomeCard/homeCard.component.js
@@ -1,17 +1,0 @@
-import React from "react";
-import "./homeCard.style.css";
-import "./../../cards.style.css";
-
-
-
-
-export const MainContent = ({children=""}) => {
-    return (
-        <main className="w-screen h-screen flex justify-center align-middle homeContainer">
-            <div className="mainContent-container p-3 rounded-xl flex flex-col justify-center">
-            {children}
-            </div>
-        </main>
-
-    )
-}
