--- client/src/composants/Cards/Large/HomeCard/homeCard.style.css
@@ -1,8 +1,0 @@
-.mainContent-container {
-    width: 80%;
-    height: 70%;
-    margin-top: 13rem;
-    background-color: var(--main-xDark);
-    -webkit-box-shadow: 7px 5px 12px 3px var(--black-0); 
-box-shadow: 7px 5px 12px 3px var(--black-0);
-}
