--- client/src/composants/Cards/Small/LinkCards/smallCard_link.js
@@ -1,6 +1,0 @@
-import React from "react";
-import "./../../cards.style.css"; // Nécessaire pour chaque card !
-import "./smallCards_link.style.css";
-
-export const CardLink =() => {
-}
