+++ client/src/composants/Discover/Artistes/Details/ArtistCardDetails.component.jsx
@@ -1,0 +1,7 @@
+export const ArtistCardDetails = () => {
+    return (
+        <>
+        
+        </>
+    )
+}
