+++ client/src/composants/Discover/Discover.component.jsx
@@ -1,0 +1,8 @@
+
+
+export const Discover = () => {
+    return (
+        <>
+        </>
+    )
+}
