+++ client/src/composants/Discover/Genres/ListGenres.component.jsx
@@ -1,0 +1,6 @@
+export const ListGenres = () => {
+    return (
+        <>
+        </>
+    )
+}
