+++ client/src/composants/Info-pratiques/Alertes/Alertes.component.jsx
@@ -1,0 +1,6 @@
+export const Alertes = () => {
+    return (
+        <>
+        </>
+    )
+}
