+++ client/src/composants/Info-pratiques/Contact/Contact.component.jsx
@@ -1,0 +1,6 @@
+export const Contact = () => {
+    return(
+        <>
+        </>
+    )
+}
