+++ client/src/composants/Info-pratiques/FAQ/FAQ.component.jsx
@@ -1,0 +1,6 @@
+export const FAQ = () => {
+    return (
+        <>
+        </>
+    )
+}
