+++ client/src/composants/Info-pratiques/Partenaires/Partenaires.component.jsx
@@ -1,0 +1,6 @@
+export const Partenaires = () => {
+    return (
+        <>
+        </>
+    )
+}
