+++ client/src/composants/Info-pratiques/infoPratique.component.jsx
@@ -1,0 +1,6 @@
+export const InfoPratiques = () => {
+    return (
+        <>
+        </>
+    )
+}
