+++ client/src/composants/Permanent/homeCard.style.css
@@ -1,0 +1,16 @@
+.mainContent-container {
+    width: 80%;
+    height: 70%;
+    margin-top: 13rem;
+    background-color: var(--main-xDark);
+    -webkit-box-shadow: 7px 5px 12px 3px var(--black-0); 
+box-shadow: 7px 5px 12px 3px var(--black-0);
+}
+
+.card {
+    display: flex;
+    flex-direction: column;
+    border: solid 2px;
+    border-radius: 0 2rem 2rem 0;
+    width: 100%;
+}
